Moved actions-interfaces to separate package
